requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.1
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-db-connection-pool==1.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.55.3
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
lxml==5.3.0
mysqlclient==2.2.7
numpy==2.3.5
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pye57==0.4.18
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pypdf==5.1.0
pyphen==0.17.0
pyquaternion==0.9.9
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
scipy==1.16.3
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.40
sqlparams==6.2.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
uvicorn==0.34.0
vercel==0.2.1
waitress==3.0.2
weasyprint==63.1
webencodings==0.5.1
wfastcgi==3.0.0
wheel==0.46.3
whitenoise==6.8.2
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
